requirements.txt
gradio
openai
langchain
weaviate-client
requests
python-dotenv
tiktoken
sentence-transformers
zhipuai
